requirements.txt
# VendoMini Core Dependencies
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0

# Optional dependencies (install separately if needed):
# pip install ray[default]>=2.7.0  # For cluster execution (Linux/Mac only)
# pip install openai>=1.0.0        # For OpenAI models
# pip install anthropic>=0.7.0     # For Anthropic models
# pip install lifelines>=0.27.0    # For survival analysis
